+++ tests/service/privacy_request/request_runner_service_test.py
@@ -38,49 +38,69 @@ from fidesops.service.masking.strategy.masking_strategy_factory import (
     MaskingStrategyFactory,
 )
 from fidesops.service.masking.strategy.masking_strategy_hmac import HmacMaskingStrategy
-from fidesops.service.privacy_request.request_runner_service import PrivacyRequestRunner
-from fidesops.util.async_util import wait_for
+from fidesops.service.privacy_request.request_runner_service import (
+    run_privacy_request,
+    run_webhooks_and_report_status,
+)
 from fidesops.util.data_category import DataCategory
+
+PRIVACY_REQUEST_TASK_TIMEOUT = 2
+# External services take much longer to return
+PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL = 15
 
 
 @mock.patch("fidesops.service.privacy_request.request_runner_service.upload")
 def test_policy_upload_called(
     upload_mock: Mock,
-    db: Session,
-    privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
+    privacy_request_status_pending: PrivacyRequest,
+    run_privacy_request_task,
 ) -> None:
-    wait_for(privacy_request_runner.submit())
+    run_privacy_request_task.delay(privacy_request_status_pending.id).get(
+        timeout=PRIVACY_REQUEST_TASK_TIMEOUT
+    )
     assert upload_mock.called
 
 
 def test_start_processing_sets_started_processing_at(
     db: Session,
-    privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
+    privacy_request_status_pending: PrivacyRequest,
+    run_privacy_request_task,
 ) -> None:
-    privacy_request.started_processing_at = None
-    wait_for(privacy_request_runner.submit())
-
+    assert privacy_request_status_pending.started_processing_at is None
+    run_privacy_request_task.delay(privacy_request_status_pending.id).get(
+        timeout=PRIVACY_REQUEST_TASK_TIMEOUT
+    )
     _sessionmaker = get_db_session()
     db = _sessionmaker()
-    privacy_request = PrivacyRequest.get(db=db, id=privacy_request.id)
-    assert privacy_request.started_processing_at is not None
+    assert (
+        PrivacyRequest.get(
+            db=db, id=privacy_request_status_pending.id
+        ).started_processing_at
+        is not None
+    )
 
 
 def test_start_processing_doesnt_overwrite_started_processing_at(
     db: Session,
     privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
+    run_privacy_request_task,
 ) -> None:
     before = privacy_request.started_processing_at
-    wait_for(privacy_request_runner.submit())
+    assert before is not None
+
+    run_privacy_request_task.delay(privacy_request.id).get(
+        timeout=PRIVACY_REQUEST_TASK_TIMEOUT
+    )
+
+    _sessionmaker = get_db_session()
+    db = _sessionmaker()
+
     privacy_request = PrivacyRequest.get(db=db, id=privacy_request.id)
     assert privacy_request.started_processing_at == before
 
 
 @mock.patch(
-    "fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run_webhooks_and_report_status",
+    "fidesops.service.privacy_request.request_runner_service.run_webhooks_and_report_status",
 )
 @mock.patch(
     "fidesops.service.privacy_request.request_runner_service.run_access_request"
@@ -92,15 +112,17 @@ def test_from_graph_resume_does_not_run_pre_webhooks(
     run_webhooks,
     db: Session,
     privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
+    run_privacy_request_task,
     erasure_policy,
 ) -> None:
     privacy_request.started_processing_at = None
     privacy_request.policy = erasure_policy
     privacy_request.save(db)
 
-    privacy_request.started_processing_at = None
-    wait_for(privacy_request_runner.submit(from_step=PausedStep.access))
+    run_privacy_request_task.delay(
+        privacy_request_id=privacy_request.id,
+        from_step=PausedStep.access.value,
+    ).get(timeout=PRIVACY_REQUEST_TASK_TIMEOUT)
 
     _sessionmaker = get_db_session()
     db = _sessionmaker()
@@ -117,7 +139,7 @@ def test_from_graph_resume_does_not_run_pre_webhooks(
 
 
 @mock.patch(
-    "fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run_webhooks_and_report_status",
+    "fidesops.service.privacy_request.request_runner_service.run_webhooks_and_report_status",
 )
 @mock.patch(
     "fidesops.service.privacy_request.request_runner_service.run_access_request"
@@ -129,14 +151,17 @@ def test_resume_privacy_request_from_erasure(
     run_webhooks,
     db: Session,
     privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
+    run_privacy_request_task,
     erasure_policy,
 ) -> None:
     privacy_request.started_processing_at = None
     privacy_request.policy = erasure_policy
     privacy_request.save(db)
 
-    wait_for(privacy_request_runner.submit(from_step=PausedStep.erasure))
+    run_privacy_request_task.delay(
+        privacy_request_id=privacy_request.id,
+        from_step=PausedStep.erasure.value,
+    ).get(timeout=PRIVACY_REQUEST_TASK_TIMEOUT)
 
     _sessionmaker = get_db_session()
     db = _sessionmaker()
@@ -155,8 +180,9 @@ def test_resume_privacy_request_from_erasure(
 def get_privacy_request_results(
     db,
     policy,
-    cache,
+    run_privacy_request_task,
     privacy_request_data: Dict[str, Any],
+    task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT,
 ) -> PrivacyRequest:
     """Utility method to run a privacy request and return results after waiting for
     the returned future."""
@@ -198,11 +224,8 @@ def get_privacy_request_results(
             for masking_secret in masking_secrets:
                 privacy_request.cache_masking_secret(masking_secret)
 
-    wait_for(
-        PrivacyRequestRunner(
-            cache=cache,
-            privacy_request=privacy_request,
-        ).submit()
+    run_privacy_request_task.delay(privacy_request.id).get(
+        timeout=task_timeout,
     )
 
     return PrivacyRequest.get(db=db, id=privacy_request.id)
@@ -220,6 +243,7 @@ def test_create_and_process_access_request(
     policy,
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
+    run_privacy_request_task,
 ):
     customer_email = "customer-1@example.com"
     data = {
@@ -228,7 +252,12 @@ def test_create_and_process_access_request(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+    )
 
     results = pr.get_results()
     assert len(results.keys()) == 11
@@ -271,6 +300,7 @@ def test_create_and_process_access_request_mssql(
     policy,
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
+    run_privacy_request_task,
 ):
 
     customer_email = "customer-1@example.com"
@@ -280,7 +310,12 @@ def test_create_and_process_access_request_mssql(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+    )
 
     results = pr.get_results()
     assert len(results.keys()) == 11
@@ -311,6 +346,7 @@ def test_create_and_process_access_request_mysql(
     policy,
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
+    run_privacy_request_task,
 ):
 
     customer_email = "customer-1@example.com"
@@ -320,7 +356,12 @@ def test_create_and_process_access_request_mysql(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+    )
 
     results = pr.get_results()
     assert len(results.keys()) == 11
@@ -352,6 +393,7 @@ def test_create_and_process_access_request_mariadb(
     policy,
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
+    run_privacy_request_task,
 ):
 
     customer_email = "customer-1@example.com"
@@ -361,7 +403,12 @@ def test_create_and_process_access_request_mariadb(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+    )
 
     results = pr.get_results()
     assert len(results.keys()) == 11
@@ -394,6 +441,7 @@ def test_create_and_process_access_request_saas_mailchimp(
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
     mailchimp_identity_email,
+    run_privacy_request_task,
 ):
     customer_email = mailchimp_identity_email
     data = {
@@ -402,7 +450,13 @@ def test_create_and_process_access_request_saas_mailchimp(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     results = pr.get_results()
     assert len(results.keys()) == 3
 
@@ -433,6 +487,7 @@ def test_create_and_process_erasure_request_saas(
     generate_auth_header,
     mailchimp_identity_email,
     reset_mailchimp_data,
+    run_privacy_request_task,
 ):
     customer_email = mailchimp_identity_email
     data = {
@@ -441,7 +496,13 @@ def test_create_and_process_erasure_request_saas(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy_hmac, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy_hmac,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
 
     connector = SaaSConnector(mailchimp_connection_config)
     request: SaaSRequestParams = SaaSRequestParams(
@@ -485,6 +546,7 @@ def test_create_and_process_access_request_saas_hubspot(
     policy_pre_execution_webhooks,
     policy_post_execution_webhooks,
     hubspot_identity_email,
+    run_privacy_request_task,
 ):
     customer_email = hubspot_identity_email
     data = {
@@ -493,7 +555,13 @@ def test_create_and_process_access_request_saas_hubspot(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     results = pr.get_results()
     assert len(results.keys()) == 3
 
@@ -521,6 +589,7 @@ def test_create_and_process_erasure_request_specific_category_postgres(
     generate_auth_header,
     erasure_policy,
     read_connection_config,
+    run_privacy_request_task,
 ):
     customer_email = "customer-1@example.com"
     customer_id = 1
@@ -533,7 +602,12 @@ def test_create_and_process_erasure_request_specific_category_postgres(
     stmt = select("*").select_from(table("customer"))
     res = postgres_integration_db.execute(stmt).all()
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -560,6 +634,7 @@ def test_create_and_process_erasure_request_specific_category_mssql(
     db,
     generate_auth_header,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = "customer-1@example.com"
     customer_id = 1
@@ -569,7 +644,12 @@ def test_create_and_process_erasure_request_specific_category_mssql(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -596,6 +676,7 @@ def test_create_and_process_erasure_request_specific_category_mysql(
     db,
     generate_auth_header,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = "customer-1@example.com"
     customer_id = 1
@@ -605,7 +686,12 @@ def test_create_and_process_erasure_request_specific_category_mysql(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -632,6 +718,7 @@ def test_create_and_process_erasure_request_specific_category_mariadb(
     db,
     generate_auth_header,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = "customer-1@example.com"
     customer_id = 1
@@ -641,7 +728,12 @@ def test_create_and_process_erasure_request_specific_category_mariadb(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -668,6 +760,7 @@ def test_create_and_process_erasure_request_generic_category(
     db,
     generate_auth_header,
     erasure_policy,
+    run_privacy_request_task,
 ):
     # It's safe to change this here since the `erasure_policy` fixture is scoped
     # at function level
@@ -683,7 +776,12 @@ def test_create_and_process_erasure_request_generic_category(
         "identity": {"email": email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -717,6 +815,7 @@ def test_create_and_process_erasure_request_aes_generic_category(
     db,
     generate_auth_header,
     erasure_policy_aes,
+    run_privacy_request_task,
 ):
     # It's safe to change this here since the `erasure_policy` fixture is scoped
     # at function level
@@ -732,7 +831,12 @@ def test_create_and_process_erasure_request_aes_generic_category(
         "identity": {"email": email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy_aes, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy_aes,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -767,6 +871,7 @@ def test_create_and_process_erasure_request_with_table_joins(
     db,
     cache,
     erasure_policy,
+    run_privacy_request_task,
 ):
     # It's safe to change this here since the `erasure_policy` fixture is scoped
     # at function level
@@ -782,7 +887,12 @@ def test_create_and_process_erasure_request_with_table_joins(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     pr.delete(db=db)
 
     stmt = select(
@@ -813,6 +923,7 @@ def test_create_and_process_erasure_request_read_access(
     db,
     cache,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = "customer-2@example.com"
     customer_id = 2
@@ -822,7 +933,12 @@ def test_create_and_process_erasure_request_read_access(
         "identity": {"email": customer_email},
     }
 
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+    )
     errored_execution_logs = pr.execution_logs.filter_by(status="error")
     assert errored_execution_logs.count() == 9
     assert (
@@ -892,6 +1008,7 @@ def test_create_and_process_access_request_snowflake(
     db,
     cache,
     policy,
+    run_privacy_request_task,
 ):
     customer_email = snowflake_resources["email"]
     customer_name = snowflake_resources["name"]
@@ -900,7 +1017,13 @@ def test_create_and_process_access_request_snowflake(
         "policy_key": policy.key,
         "identity": {"email": customer_email},
     }
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     results = pr.get_results()
     customer_table_key = (
         f"EN_{pr.id}__access_request__snowflake_example_test_dataset:customer"
@@ -921,6 +1044,7 @@ def test_create_and_process_erasure_request_snowflake(
     db,
     cache,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = snowflake_resources["email"]
     snowflake_client = snowflake_resources["client"]
@@ -930,7 +1054,13 @@ def test_create_and_process_erasure_request_snowflake(
         "policy_key": erasure_policy.key,
         "identity": {"email": customer_email},
     }
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     pr.delete(db=db)
 
     stmt = f'select "name", "variant_eg" from "customer" where "email" = {formatted_customer_email};'
@@ -996,10 +1126,7 @@ def redshift_resources(
 @pytest.mark.integration_external
 @pytest.mark.integration_redshift
 def test_create_and_process_access_request_redshift(
-    redshift_resources,
-    db,
-    cache,
-    policy,
+    redshift_resources, db, cache, policy, run_privacy_request_task
 ):
     customer_email = redshift_resources["email"]
     customer_name = redshift_resources["name"]
@@ -1008,7 +1135,13 @@ def test_create_and_process_access_request_redshift(
         "policy_key": policy.key,
         "identity": {"email": customer_email},
     }
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     results = pr.get_results()
     customer_table_key = (
         f"EN_{pr.id}__access_request__redshift_example_test_dataset:customer"
@@ -1039,6 +1172,7 @@ def test_create_and_process_erasure_request_redshift(
     db,
     cache,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = redshift_resources["email"]
     data = {
@@ -1048,7 +1182,13 @@ def test_create_and_process_erasure_request_redshift(
     }
 
     # Should erase customer name
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     pr.delete(db=db)
 
     connector = redshift_resources["connector"]
@@ -1073,7 +1213,13 @@ def test_create_and_process_erasure_request_redshift(
     target.save(db=db)
 
     # Should erase state fields on address table
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     pr.delete(db=db)
 
     connector = redshift_resources["connector"]
@@ -1097,6 +1243,7 @@ def test_create_and_process_access_request_bigquery(
     db,
     cache,
     policy,
+    run_privacy_request_task,
 ):
     customer_email = bigquery_resources["email"]
     customer_name = bigquery_resources["name"]
@@ -1105,7 +1252,13 @@ def test_create_and_process_access_request_bigquery(
         "policy_key": policy.key,
         "identity": {"email": customer_email},
     }
-    pr = get_privacy_request_results(db, policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     results = pr.get_results()
     customer_table_key = (
         f"EN_{pr.id}__access_request__bigquery_example_test_dataset:customer"
@@ -1136,6 +1289,7 @@ def test_create_and_process_erasure_request_bigquery(
     db,
     cache,
     erasure_policy,
+    run_privacy_request_task,
 ):
     customer_email = bigquery_resources["email"]
     data = {
@@ -1145,7 +1299,13 @@ def test_create_and_process_erasure_request_bigquery(
     }
 
     # Should erase customer name
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     pr.delete(db=db)
 
     bigquery_client = bigquery_resources["client"]
@@ -1168,7 +1328,13 @@ def test_create_and_process_erasure_request_bigquery(
     target.save(db=db)
 
     # Should erase state fields on address table
-    pr = get_privacy_request_results(db, erasure_policy, cache, data)
+    pr = get_privacy_request_results(
+        db,
+        erasure_policy,
+        run_privacy_request_task,
+        data,
+        task_timeout=PRIVACY_REQUEST_TASK_TIMEOUT_EXTERNAL,
+    )
     pr.delete(db=db)
 
     bigquery_client = bigquery_resources["client"]
@@ -1183,23 +1349,20 @@ def test_create_and_process_erasure_request_bigquery(
             assert row.state is None
 
 
-class TestPrivacyRequestRunnerRunWebhooks:
+class TestRunPrivacyRequestRunsWebhooks:
     @mock.patch("fidesops.models.privacy_request.PrivacyRequest.trigger_policy_webhook")
     def test_run_webhooks_halt_received(
         self,
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
         mock_trigger_policy_webhook.side_effect = PrivacyRequestPaused(
             "Request received to halt"
         )
 
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
-            db, privacy_request, PolicyPreWebhook
-        )
+        proceed = run_webhooks_and_report_status(db, privacy_request, PolicyPreWebhook)
         assert not proceed
         assert privacy_request.finished_processing_at is None
         assert privacy_request.status == PrivacyRequestStatus.paused
@@ -1211,7 +1374,6 @@ class TestPrivacyRequestRunnerRunWebhooks:
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
         config.redis.DEFAULT_TTL_SECONDS = (
@@ -1221,9 +1383,7 @@ class TestPrivacyRequestRunnerRunWebhooks:
             "Request received to halt"
         )
 
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
-            db, privacy_request, PolicyPreWebhook
-        )
+        proceed = run_webhooks_and_report_status(db, privacy_request, PolicyPreWebhook)
         assert not proceed
         time.sleep(3)
 
@@ -1239,16 +1399,13 @@ class TestPrivacyRequestRunnerRunWebhooks:
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
         mock_trigger_policy_webhook.side_effect = ClientUnsuccessfulException(
             status_code=500
         )
 
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
-            db, privacy_request, PolicyPreWebhook
-        )
+        proceed = run_webhooks_and_report_status(db, privacy_request, PolicyPreWebhook)
         assert not proceed
         assert privacy_request.status == PrivacyRequestStatus.error
         assert privacy_request.finished_processing_at is not None
@@ -1260,16 +1417,13 @@ class TestPrivacyRequestRunnerRunWebhooks:
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
         mock_trigger_policy_webhook.side_effect = ValidationError(
             errors={}, model=SecondPartyResponseFormat
         )
 
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
-            db, privacy_request, PolicyPreWebhook
-        )
+        proceed = run_webhooks_and_report_status(db, privacy_request, PolicyPreWebhook)
         assert not proceed
         assert privacy_request.finished_processing_at is not None
         assert privacy_request.status == PrivacyRequestStatus.error
@@ -1281,13 +1435,10 @@ class TestPrivacyRequestRunnerRunWebhooks:
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
 
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
-            db, privacy_request, PolicyPreWebhook
-        )
+        proceed = run_webhooks_and_report_status(db, privacy_request, PolicyPreWebhook)
         assert proceed
         assert privacy_request.status == PrivacyRequestStatus.in_processing
         assert privacy_request.finished_processing_at is None
@@ -1299,11 +1450,10 @@ class TestPrivacyRequestRunnerRunWebhooks:
         mock_trigger_policy_webhook,
         db,
         privacy_request,
-        privacy_request_runner,
         policy_pre_execution_webhooks,
     ):
         """Test running webhooks after specific webhook - for when we're resuming privacy request execution"""
-        proceed = privacy_request_runner.run_webhooks_and_report_status(
+        proceed = run_webhooks_and_report_status(
             db, privacy_request, PolicyPreWebhook, policy_pre_execution_webhooks[0].id
         )
         assert proceed