+++ tests/tasks/test_celery.py
@@ -1,10 +1,10 @@
-def test_create_task(celery_app, celery_worker):
-    @celery_app.task
+def test_create_task(celery_session_app, celery_session_worker):
+    @celery_session_app.task
     def multiply(x, y):
         return x * y
 
     # Force `celery_app` to register our new task
     # See: https://github.com/celery/celery/issues/3642#issuecomment-369057682
-    celery_worker.reload()
+    celery_session_worker.reload()
     assert multiply.run(4, 4) == 16
     assert multiply.delay(4, 4).get(timeout=10) == 16